// src/types/enum.types.ts
export enum ROLE {
  USER = "USER",
  ADMIN = "ADMIN",
  SUPER_ADMIN = "SUPER_ADMIN",
}

export enum ERROR_CODES {
  VALIDATION_ERR = "VALIDATION_ERR",
  NOT_FOUND_ERR = "NOT_FOUND_ERR",
  INTERNAL_SERVER_ERR = "INTERNAL_SERVER_ERR",
  AUTH_ERR = "AUTH_ERR",
}

export const OnlyAdmins = [ROLE.ADMIN,ROLE.SUPER_ADMIN]
export const OnlyUsers = [ROLE.USER]